Add EventPage component for event listing with carousel functionality; implement AOS for animations and enhance event detail page with structured data retrieval.
